requirements.txt
ansible==9.13.0
ansible-core==2.16.14
argcomplete==3.6.0
azure-core==1.30.2
azure-storage-blob==12.20.0
b2sdk==1.21.0
bcrypt==4.1.2
Beaker==1.12.1
beautifulsoup4==4.13.3
boto3==1.37.33
botocore==1.37.33
Brlapi==0.8.5
cachetools==5.4.0
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
configobj==5.0.8
cryptography==43.0.0
cupshelpers==1.0
dbus-python==1.3.2
debtcollector==3.0.0
distlib==0.3.8
distro==1.9.0
dnf==4.23.0
dropbox==12.0.2
duplicity==3.0.4.1
fasteners==0.19
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
fros==1.1
google-api-core==2.11.1
google-api-python-client==2.140.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.63.0
grpcio==1.71.0
grpcio-status==1.48.4
grpcio-tools==1.71.0
httplib2==0.21.0
humanize==4.9.0
identify==2.6.9
idna==3.7
iso8601==1.1.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.6
jmespath==1.0.1
keyring==25.6.0
keystoneauth1==5.6.0
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
libdnf5==5.2.12.0
libvirt-python==10.6.0
logfury==1.0.1
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
MarkupSafe==2.1.5
more-itertools==10.1.0
msgpack==1.0.6
netaddr==1.3.0
netifaces==0.11.0
nftables==0.1
nodeenv==1.9.1
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.47
os-service-types==1.7.0
oslo.config==9.6.0
oslo.i18n==6.3.0
oslo.serialization==5.4.0
oslo.utils==7.1.0
packaging==24.2
paramiko==3.5.1
Paste==3.10.1
pbr==5.11.1
perf==0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
ply==3.11
podman-compose==1.3.0
pre_commit==3.8.0
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
PyDrive2==1.20.0
pyenchant==3.2.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-keystoneclient==5.4.0
python-linux-procfs==0.7.3
python-pam==2.0.2
python-swiftclient==4.5.0
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
ramalama==0.7.3
receptorctl==1.5.3
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
resolvelib==1.0.1
rfc3986==1.5.0
rh-gitleaks==2.3.2
rh-pre-commit==2.3.2
rpm==4.20.1
rsa==4.9
s3transfer==0.11.4
SecretStorage==3.3.3
setools==4.5.1
setuptools==69.2.0
shtab==1.7.1
simpleaudio==1.0.4
six==1.16.0
sos==4.8.2
soupsieve==2.6
SSSDConfig==2.10.2
stevedore==5.2.0
stone==3.2.1
systemd-python==235
Tempita==0.5.2
terminator==2.1.4
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
virtualenv==20.21.1
wrapt==1.17.1
mcp==1.5.0
uv==0.6.10